+++ example/lib/screens/component/apple_pay/apple_pay_advanced_component_screen.dart
@@ -48,8 +48,8 @@ class ApplePayAdvancedComponentScreen extends StatelessWidget {
       future: repository.fetchPaymentMethods(),
       builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
         if (snapshot.hasData) {
-          final AdvancedCheckoutPreview advancedCheckout =
-              AdvancedCheckoutPreview(
+          final AdvancedCheckout advancedCheckout =
+              AdvancedCheckout(
                   onSubmit: repository.onSubmit,
                   onAdditionalDetails: repository.onAdditionalDetailsMock);
           final paymentMethod = _extractPaymentMethod(snapshot.data!);