+++ example/lib/screens/drop_in/drop_in_screen.dart
@@ -74,7 +74,7 @@ class _DropInScreenState extends State<DropInScreen> {
     final paymentMethodsResponse =
         await widget.repository.fetchPaymentMethods();
     final dropInConfiguration = await _createDropInConfiguration();
-    final advancedCheckoutPreview = AdvancedCheckoutPreview(
+    final advancedCheckout = AdvancedCheckout(
       onSubmit: widget.repository.onSubmit,
       onAdditionalDetails: widget.repository.onAdditionalDetails,
     );
@@ -82,7 +82,7 @@ class _DropInScreenState extends State<DropInScreen> {
     final paymentResult = await AdyenCheckout.advanced.startDropIn(
       dropInConfiguration: dropInConfiguration,
       paymentMethodsResponse: paymentMethodsResponse,
-      advancedCheckoutPreview: advancedCheckoutPreview,
+      checkout: advancedCheckout,
     );
 
     _showPaymentResultDialog(paymentResult);