+++ lib/src/components/instant/instant_advanced_component.dart
@@ -8,7 +8,7 @@ import 'package:adyen_checkout/src/util/constants.dart';
 import 'package:adyen_checkout/src/util/payment_event_handler.dart';
 
 class InstantAdvancedComponent extends BaseInstantComponent {
-  final AdvancedCheckoutPreview advancedCheckout;
+  final AdvancedCheckout advancedCheckout;
   final PaymentEventHandler paymentEventHandler;
 
   @override