+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Updated Android SDK to v5.3.1.
 * Added support for customizing Drop-in and card component.
 * Added option to adjust Drop-in preselected payment method title.
+* Replaced advanced checkout class with the preview implementation.
 
 ## 0.0.3